# docs/swagger.yaml
definitions:
  entity.DictionaryEntry:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
  entity.DictionaryEntryDTO:
    properties:
      name:
        type: string
    type: object
info:
  contact: {}
paths:
  /category:
    get:
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.DictionaryEntry'
            type: array
      tags:
      - категория
    post:
      parameters:
      - description: категория
        in: body
        name: dto
        required: true
        schema:
          $ref: '#/definitions/entity.DictionaryEntryDTO'
      responses: {}
      tags:
      - категория
  /category/{id}:
    delete:
      parameters:
      - description: идентификатор
        in: path
        name: id
        required: true
        type: string
      responses: {}
      tags:
      - категория
    patch:
      parameters:
      - description: категория
        in: body
        name: dto
        required: true
        schema:
          $ref: '#/definitions/entity.DictionaryEntryDTO'
      - description: идентификатор
        in: path
        name: id
        required: true
        type: string
      responses: {}
      tags:
      - категория
  /country:
    get:
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.DictionaryEntry'
            type: array
      tags:
      - страна
    post:
      parameters:
      - description: страна
        in: body
        name: dto
        required: true
        schema:
          $ref: '#/definitions/entity.DictionaryEntryDTO'
      responses: {}
      tags:
      - страна
  /country/{id}:
    delete:
      parameters:
      - description: идентификатор
        in: path
        name: id
        required: true
        type: string
      responses: {}
      tags:
      - страна
    patch:
      parameters:
      - description: страна
        in: body
        name: dto
        required: true
        schema:
          $ref: '#/definitions/entity.DictionaryEntryDTO'
      - description: идентификатор
        in: path
        name: id
        required: true
        type: string
      responses: {}
      tags:
      - страна
swagger: "2.0"